@@ -22,8 +22,8 @@
     "dependencies": {
         "@react-spring/web": "^9.4.4",
         "lodash": "^4.17.21",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
+        "react": "^18.0.0",
+        "react-dom": "^18.0.0",
         "styled-components": "^5.3.5",
         "styled-normalize": "^8.0.7"
     },
@@ -31,8 +31,8 @@
         "@parcel/config-default": "^2.4.1",
         "@parcel/transformer-svg-react": "^2.4.1",
         "@types/lodash": "^4.14.182",
-        "@types/react": "^17.0.44",
-        "@types/react-dom": "^17.0.15",
+        "@types/react": "^18.0.5",
+        "@types/react-dom": "^18.0.1",
         "@types/styled-components": "^5.1.25",
         "@typescript-eslint/eslint-plugin": "^5.20.0",
         "@typescript-eslint/parser": "^5.20.0",
@@ -44,5 +44,9 @@
         "parcel": "^2.4.1",
         "prettier": "^2.6.2",
         "typescript": "^4.6.3"
+    },
+    "resolutions": {
+        "**/@types/react": "^18.0.5",
+        "**/@types/react-dom": "^18.0.1"
     }
 }
